kbs: Move the Confidential Containers AS to its own crate
Allow for supporting other attestation services.
