+++ src/api/src/attestation/coco/grpc.rs
@@ -2,7 +2,7 @@
 // Licensed under the Apache License, Version 2.0, see LICENSE for details.
 // SPDX-License-Identifier: Apache-2.0
 
-use super::Attest;
+use crate::attestation::Attest;
 use crate::config::Config;
 use anyhow::*;
 use as_types::AttestationResults;
